gateway: extract _files and _images
